TP2 : déplacement des personnages + affichage des RTT avec actualisation du RTT en temps réel

+++ network_core/includes/messages/gameplay_message.hpp
@@ -6,14 +6,14 @@
 class GameplayMessage : public BaseMessage {
 public:
     GameplayMessage();
-    GameplayMessage(NetworkID network_id, ClassID class_id, uint8_t x, uint8_t y);
+    GameplayMessage(NetworkID network_id, ClassID class_id, uint32_t x, uint32_t y);
 
     NetworkID _network_id;
     ClassID _class_id;
 
     // Initial player position
-    uint8_t _x;
-    uint8_t _y;
+    uint32_t _x;
+    uint32_t _y;
 
 protected:
     virtual void SerializePayload() const override;
